Demonstrate the AS-Unsafe deadlock when any thread holds the lock
A mutex is the mechanism typically used to avoid race-conditions in
multi-threaded applications. In such applications, each thread will try
to lock a mutex, thus obtaining exclusive access to a critical region
(usually some small portion of the code), before actually accessing it,
and will refrain from doing so if some other thread is holding the lock.
Usually, when the lock operation cannot happen, the thread will wait for
the mutex to be released. An MT-Safe function is one that does not have
a critical section at all, or one that protects critical sections with
mutexes.

If an MT-Safe function does use mutexes, then it is not safe to call
from a signal handler, and it is said to be AS-Unsafe. The danger comes
from the fact that asynchronous signals can be delivered to the
application at any moment, including when one of its threads is holding
the lock to a mutex. If that thread gets selected to handle the signal
and it calls the AS-Unsafe function, it will try to lock the mutex
again, so it will wait on itself, thus causing a deadlock.

This deadlock only occurs in this specific situation, i.e. when the
thread that gets selected to handle the signal is the one holding the
lock. Conversely, if some other thread was holding the lock, the
deadlock would not arise, as the other threads would be free to resume
their execution, eventually releasing the lock, and letting the signal
handler take it.

However, on libpulp, the chances for deadlock are higher, because all
threads are stopped and put into busy waiting (thread hijacking) during
the whole duration of the signal handling, so any thread holding the
lock leads to a deadlock. This patch moves the tight malloc/free loop
from the main thread (which is always selected to install the live
patch) to all other threads in the test program, which demonstrate the
scenario described above.