Use Local Dynamic TLS Model in trm.S
ELF Handling For Thread-Local Storage [1] describes several models for
the access of thread-local variables. When thread-local variables are
defined in the same object they are referenced in, the Local Dynamic TLS
Model can be used. This model is only useful if more than one variable
is accessed, because the optimization relies on the fact that a single
call to __tls_get_addr can be made, and the result can be used in simple
additions to determine the address of the multiple variables.

In trm.S, two thread-local variables, __ulp_thread_universe and
__ulp_ret, are defined and accessed, so the Local Dynamic TLS Model can
be used. As a matter of fact, it is already partially used, because the
argument to __tls_get_addr is obtained with:

    leaq <var name>@tlsld(%rip), %rdi
    (notice the use of '@tlsld', not '@tlsgd')

As such, __tls_get_addr need not be called twice. This patch removes the
redundant call, which has a slight impact on performance (running
openssl's speed command, with cypher set to aes-128-cbc, is ~2% faster
after this patch).

[1] http://www.akkadia.org/drepper/tls.pdf