Relicense under LGPL version 2.1 or later
GPL version 3 is incompatible with common free software licenses, such
as with the OpenSSL license [1]. This raises concerns about the
legality of linking trm.o with libcrypto or libssl, even though
libpulp.so is never directly linked against them, but LD_PRELOAD'ed.

This patch relicenses the whole of libpulp under LGPL version 2.1 or, at
the option of redistributors, any later version. Alternatively,
relicensing only trm.S could fix the compatibility problem.

This patch was created by a set of scripts, which, apart from changing
the license, update copyright years and add copyright notices to files
that lacked it. The license file was downloaded from gnu.org [2].

[1] https://www.gnu.org/licenses/license-list.en.html#OpenSSL
[2] https://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt