Add test to demonstrate AS-Unsafe conversion
A program may call AS-Safe functions in signal handlers, however, if
that function gets live-patched, its prologue will make a call into
libulp's universe handling routine, which calls dlsym, thus rendering
the live-patched function AS-Unsafe. This means that the live-patched
function can no longer be called from the signal handler, otherwise it
might deadlock. This patch adds a test case that demonstrates this
problem.

Since the problem happens because of a call to dlsym, the test case
relies on the repetitive use of functions from the dlopen family.
However, this might cause an earlier bug to be exposed, which is more
frequently detected by another test case (tests/deadlock.py). When that
happens, this new test case returns a SKIP code, because it was unable
to detect the AS-Unsafe conversion.