Delay live patching in examples
The examples, as well as the test suite, use scripts to start live
patchable process, then call the trigger tool to live patch them. The
trigger tool relies on the /proc filesystem to obtain required
information about the process, such as its binary file, threads, auxv.

Right after starting the process, /proc/<pid>/exe points to
/usr/bin/bash or /usr/bin/python3, rather than to the target binary
subsequently loaded by them. Afterwards, the link gets updated, probably
after bash or python call execve. The update is quick, but not as quick
as the scripts in the test suite, which causes errors similar to the
following during make check:

  ulp: to be patched object (examples/dummy/.libs/libdummy.so.0) not loaded.
  Attempt 0 to apply live patch #1 failed (unknown error (5)). Try again later.
  ulp: read_memory error.
  ulp: error reading link_map address.
  ulp: PTRACE_ATTACH error.
  ulp: Unable to attach to 1942.

This problem is unlikely to happen in real-world scenarios, because
starting a process then immediately live patching them would be better
solved with a regular binary update. On the other hand, the test suite
starts the process just to live patch it, so, to work around the
problem, the wrapper script in examples/dummy/ now waits for the
symbolic link to be updated. The wrapper scripts in tests/ do not
require the change, because they already wait for process readiness.