Add test case (XFAIL) for local symbols live patching
Live patching LOCAL symbols is currently impossible, because libpulp
relies on the .dynsym to find the addresses of symbols it must patch.
This patch adds a test case that demonstrate the problem.

It should be possible to change the current behavior and use the .symtab
with libraries that have not been stripped. Furthermore, since LOCAL
symbols play no role in library entrance tracking, it should be safe
(from a consistency point of view) to patch them.