Avoid the red zone
Even though libpulp does not register signal handlers with the kernel,
it uses ptrace to hijack all threads in a process, then diverts the
execution of one of these threads to apply live patches and check
patching status. Thus, live patching happens from a context similar to
that of signal handlers, therefore, it must follow the rules of the ABI,
more specifically, it cannot touch the red zone.

With regular signal handlers, the Linux kernel adjusts the stack pointer
before transferring control to registered handlers. Since libpulp uses
ptrace and thread hijacking, instead of regular handler registering, it
cannot rely on this kernel feature, so it must adjust the stack on its
own. This patch implements the adjustment.