Do not mess with the stack in ulp_interface.S
During thread hijacking, the trigger tool saves the current machine
state, including the stack pointer, from every thread in the process.
Afterwards, when calling any of the functions in the critical section,
the trigger tool always uses the values it saved during hijacking, and
never relies on what happened during the execution of these function.
Thus, saving and restoring the stack pointer is not necessary in
ulp_interface.S.

Besides that, pushing the stack address onto the stack itself, modifying
its value so that it becomes aligned on 16-bytes boundaries, then
popping it would not work as intended.

This patch removes the unnecessary pushes and pops, as well as it
removes the comment about optimizing __ulp_get_path_buffer, because
these functions are not part of hot paths.