Demonstrate red zone overwriting
Live patching happens from the context of a signal handler, thus, as
mandated by the AMD64 ABI, it should not modify the contents of the red
zone, i.e. the 128-bytes long area below the stack pointer. In typical
uses of signal handlers, avoiding the red zone is trivial, because the
Linux kernel adjusts the stack pointer before actually transferring
control to registered signal handlers. However, libpulp does not
register signal handlers with the kernel, instead, it uses ptrace to
hijack the threads, hence it should adjust the stack pointer on its on.

Currently, libpulp does not adjust the stack pointer, so it violates the
ABI and writes to the red zone. This patch adds a test case that
demonstrates the problem.