Extend the integer list of parameters in the test suite
According to the ABI for x86_64, the first to sixth parameters of the
POINTER or INTEGER classes are passed on registers (%rdi, %rsi, %rdx,
%rcx, %r8, and %r9), whereas the seventh and subsequent parameters of
the same class are passed on the stack. However, the function in the
parameters test case only had four integer parameters, thus leaving %r8
and %r9 untested. This patch extends the number of integer parameters to
eight and fixes the coverage, as well as it replaces the use of signals
with standard input, removing unnecessary complexity.