Add float parameters to the test suite
According to the ABI for x86_64, the first to eigth parameters of the
SSE class are passed on vector registers (%xmm0 to %xmm7). Likewise,
parameters that fall into the SSEUP class use the same set of registers.
This patch adds a new function to the test suites, to cover this aspect
of the ABI.