Test patching of functions near page boundaries
A live patch is composed of a DSO and a metadata file. The metadata file
contains information about the to-be-patched library and a list of
to-be-patched functions. During the application of the live patch,
libpulp iterates over this list, adding control information to internal
data structures (detours), and patching the function prologue.

However, when a function resides near a memory page boundary, the live
patching process fails silently, and libpulp looses control of the
target process, thus stopping the patching process altogether, and
leaving the remaining functions in the list unattended.

This patch adds a test cases that demonstrates the problem.