Do not SIGSTOP nor SIGCONT processes during live patching
As demonstrated by the test suite, when live patching sends SIGSTOP and
SIGCONT to the target process, it might unnecessarily disturb the parent
process, for instance, it might make the terminal release from the wait,
as if Ctrl+z had been issued.

This patch refactors the way libpulp hijacks a process. Instead of
stopping (SIGSTOP) the whole process, attaching (with ptrace) to all
threads, putting them into infinite loops, detaching from all, and
finally restarting the process (SIGCONT); now it simply attaches to all
threads and keeps them attached, except for the main thread, which is
actually hijacked to execute code from libpulp. When the live patching
operations are over, the main thread context is restored, and all
threads (including the main one) get detached from, which resumes the
process.

This new method removes the need to stop and restart the process with
SIGSTOP and SIGCONT, which solves the terminal release problem. Besides
that, it unintentionally solves the problem that libpulp's tools fail to
attach to the target process, hence the replacing of the 77 (SKIP)
return code, with 99 (HARD ERROR) in terminal.py.