Test for terminal release
When a live patchable process gets started from a terminal, the terminal
hangs, as usual, until the process finishes or gets killed/terminated.
However, when applying live patches, libpulp's tools send SIGSTOP to the
target process, which may cause the terminal to release, which is not a
common behavior and might mislead users into believing that the process
got killed. This patch adds a test case that mimics this behavior.

Sometimes, and due to a different problem, libpulp's check tool fails to
attach to the target process. When that happens, the test case exits
with the SKIP error code; a separate test case is needed for this
specific error.