Fix call frame information
Library entrance tracking needs to save and restore lots of registers on
the stack. Apart from that, the trampolines in the .ulp section (emitted
by dynsym_gate) push the address of the target function onto the stack,
as well. However, none of these push and pop operations have been coded
as CFI directives for GDB, which makes backtracing unreliable.

Taking tests/blocked.c as example, if libblocked had not been made live
patchable, the backtrace of a call to hello via hello_loop would look
similar to the following in gdb:

  #0  0x00007f3df081f13f in hello () from libblocked.so.0
  #1  0x00007f3df081f17a in hello_loop () at tests/libblocked.c:37
  #2  0x0000000000401215 in function2 (arg=0x0) at tests/blocked.c:59
  #3  0x00007f3df07efeb1 in start_thread (arg=0x7f3defe14640) ...
  #4  0x00007f3df071dccf in clone () ...

However, if libblocked is made live patchable, going into the library
means going through a trampoline and through __ulp_entry. However, since
__ulp_entry does not fix CFI information, the backtrace becomes
unreliable and the call stack shows functions that normally do no show
up, for instance:

  #0  __ulp_entry () at trm.S:48
  #1  0x00007fa96926514f in frame_dummy () from libblocked.so.0
  #2  0x00007fa96926518a in hello_loop () at libblocked.c:37
  #3  0x00007fa969265248 in __ulp_thread_counter_update () at trm.S:115
  #4  0x0000000000000000 in ?? ()

Or even something more weird, such as with libdozen:

  #0  0x00007f3d72bfb18b in __ulp_entry () from libdozens.so.0
  #1  0x00000000004012a5 in __libc_csu_init (...)
  #2  0x00007f3d72a3de0a in __libc_start_main (...)
  #3  0x00000000004011aa in _start () ...

This patch fixes the CFI, which makes call stacks more reliable.
However, __ulp_entry reuses the return slot on the frame of the calling
function (so that it does not have to copy parameters around), which
causes the backtracing to stop, as can be seen below:

  #0  0x00007f9414e9714f in hello () from libblocked.so.0
  #1  0x00007f9414e9718a in hello_loop () at libblocked.c:37
  #2  0x00007f9414e97248 in __ulp_entry () at trm.S:153
  Backtrace stopped: frame did not save the PC

The return address of the calling function is saved on the TLS
(__ulp_ret), and perhaps that could be used to continue the unwinding,
but this patch only goes as far as fixing the call stack until library
entrance.

Apart from fixing CFI, this patch adds the '.L' prefix to a couple of
labels, so that they don't show up on the call frame as function names.