Convert fibonacci example into a recursion test case
This patch moves the fibonacci example from the examples directory,
where it demonstrated a worst-case scenario for the performance impact
of libpulp, to the test suite, where it now provides an example of
live-patching of recursive functions. Since recursive functions make
calls to themselves, they help test the consistency model.