Less subdirectories for the tools
According to the manual for automake, make recursion has its own serious
issues and drawbacks [1], even though recursively invoking make in
subdirectories is a supported feature.

In libpulp, each tool resides in a subdirectory, which only works if the
order of the items in the SUBDIRS variable is correctly chosen.
Otherwise, errors similar to the following could pop up:

  make[2]: *** No rule to make target '../trigger/libcommon.la', needed by 'ulp_dump'.  Stop.

On top of that, subdirectories frequently require the use of relative
paths, such as '../trigger/libcommon.la' (above) and several occurrences
of '#include "../<some header>.h)"'.

This patch moves all tools to the same directory.

[1] https://www.gnu.org/software/automake/manual/html_node/Directories.html