Fix operand size in trm.S
As correctly reported by the assembler, the uses of the cmp instruction
in trm.S do not explicitly select the size of the operands, nor can it
be determined implicitly (for instance, based on the name of register
operands):

  trm.S: Assembler messages:
  trm.S:66: Warning: no instruction mnemonic suffix given and no register operands; using default for `cmp'
  trm.S:173: Warning: no instruction mnemonic suffix given and no register operands; using default for `cmp'

The warning message states that the default for cmp is used, which,
according to the programmer's manual for x86 [1] defaults to 32-bits
operands. However, __ulp_ret holds pointers to memory, thus 64-bits
values.

When 64-bits operands are required, the instruction must be REX-prefixed
accordingly. Otherwise, only the 32 least significant bits are tested.
This is, in theory, a problem for libpulp, because __ulp_ret could point
to an address that has all 32 LSBs zeroed.

This patch replaces cmp with cmpq in trm.S, which solves the problem.

[1] AMD64 Architecture Programmer's Manual. Volume 3. Appendix B.
Section B.1 and Table B-1.