Centralize definition of padding nops amounts
Each function in a live patchable library is preceded by a pre-defined
amount of nop instructions: most nops are issued before the entry points
of these functions, and a couple of nops after it. These amounts are
used in different parts of libpulp: 1. libpulp.so uses it to know how
many bytes it needs to patch; 2. dynsym_gate, to convert multiple nops
into a single instruction; 3. the build system, as argument to gcc's
-fpatchable-function-entry.

Instead of redefining these amounts on multiple locations, this patch
adds configure rules to spread the information.