Extend the red zone test
The redzone test performs loads and stores just a few bytes away from
the stack pointer, which were capable of detecting problems when libpulp
did not make adjustments to the stack before applying live patches.

However, since the commit

commit 5be607ae1992b6cc2472ae6536a641e9e20576f2
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Fri Oct 9 14:38:07 2020 -0300

    Avoid the red zone

libpulp fixed the lack of adjustments, and now the test case is unlikely
to detect problems if libpulp were to write near the end of the redzone.
This patch extends the test case, so that it performs loads and writes
near the end of the redzone, as well. The new test did not reveal any
problems, though.