Hide symbols that need not be exported
Libpulp, the library, is not supposed to be linked against applications.
In order to apply a live patch, one should use libpulp's tools, which:
attach to running processes with ptrace; find the dynamic addresses of
required symbols (functions and objects); hijack the process; and,
finally, execute code from (or modify data in) libpulp.

The tools use ptrace to collect information about the running process,
such as the addresses where symbols have been loaded in memory. Thus,
most of the currently exported symbols need not be exported at all, even
if they are accessed from outside the library. Besides that, many
internal symbols are not declared static, thus exported, as well.

This patch adds a version script to the compilation of libpulp.so, so
that no symbols are exported unnecessarily. As a matter of fact, a
single symbol is required, __ulp_global_universe, which is accessed via
GOT from live patchable libraries.