Add explanation about syscall restart and nops
Libpulp's tools, such as ulp_trigger, use ptrace to attach to a process,
then tamper with the program counter in order to execute code from
libpulp.so. However, upon receiving PTRACE_CONT, the program counter
might be modified again, now by the kernel, so that ongoing syscalls get
restarted (when that happens, the program counter gets subtracted by
two, which is the size of the syscall instruction).

To account for this situation, libpulp's tools always set the program
counter to an address preceded by two nop instructions. If the two nops
were not present, random code, from the preceding function, could be
executed, potentially leading to bugs and crashes. The following,
step-by-step execution in gdb exemplifies the problem:

Attach to the process and check its current state:

  >>> info registers rip rsp
  rip            0x7fc6004fe222      0x7fc6004fe222 <__GI___libc_read+18>
  rsp            0x7fff938b42e8      0x7fff938b42e8

  (note: __GI___libc_read+18 is a syscall, more specifically, read)

Hijack the process (set %rip to a function from libpulp.so):

  >>> set $rip=__ulp_trigger
  >>> info registers rip rsp
  rip            0x7fc6005f4e59      0x7fc6005f4e59 <__ulp_trigger>
  rsp            0x7fff938b42e8      0x7fff938b42e8

Show what code the program counter points to (and the previous 2 bytes):

  >>> x /5i __ulp_trigger-2
     0x7fc6005f4e57 <__ulp_testlocks+6>:    dec    %esp
  => 0x7fc6005f4e59 <__ulp_trigger>:        nop
     0x7fc6005f4e5a <__ulp_trigger+1>:      nop
     0x7fc6005f4e5b <__ulp_trigger+2>:      callq  0x7fc6005f4a30 <__ulp_apply_patch>
     0x7fc6005f4e60 <__ulp_trigger+7>:      int3

Execute a single instruction and check the results:

  >>> si
  0x00007fc6005f4e59 in __ulp_trigger () from /home/gabriel/build/x86_64/libpulp/lib/.libs/libpulp.so

  >>> info registers rip rsp
  rip            0x7fc6005f4e59      0x7fc6005f4e59 <__ulp_trigger>
  rsp            0x938b42e7          0x938b42e7

Notice that, even after executing a single instruction, %rip still
points to 0x7fc6005f4e59 <__ulp_trigger>, which shows that some other
instruction was executed before. We also know that the kernel might have
subtracted two from the %rip (to restart an ongoing syscall), and we can
see above that the instruction at 0x7fc6005f4e57 <__ulp_trigger-2>
indeed decremented %esp.