Demonstrate that all threads must be restored by introspection tools
The commit

commit 32428e103112a6a54af08f848908f6904e9c5422
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Thu Oct 29 15:33:09 2020 -0300

    Do not SIGSTOP nor SIGCONT processes during live patching

removed the context restoration of non-main threads, based on the
assumption that the trigger and check tools only ever used the main
thread of a hijacked process to run code from libpulp. However,
introspection routines offer the possibility of using different threads
for that, and, as a matter of fact, the ulp tool already uses it.

This patch adds a test case that demonstrates that restoring all threads
is necessary.