Add test case for syscall restarting
When a thread that is in the middle of a syscall receives a signal, the
syscall gets interrupted. On some occasions, such as when there are no
handlers registered for the signal, or when the SA_RESTART flag is in
effect (see sigaction(2)), the kernel arranges an automatic restarting
of the syscall to happen when the thread is rescheduled, and it does so
by subtracting one instruction from the program counter.

Since attaching to processes causes a stopping signal to be sent to the
attached thread, libpulp's tools trigger this subtraction mechanism.
Even when libpulp explicitly and intentionally modifies the program
counter to somewhere completely different (to execute live patching
routines), the subtraction happens. Thus, in order to avoid execution of
stray instructions, live patching routines are preceded by some nops.

On the one hand, libpulp works correctly. On the other hand, the kernel
is unaware that libpulp tampered with the program counter, so it cannot
be sure that the syscall instruction was indeed re-executed. A problem
would happen if the kernel assumed that the program counter subtraction
always led to the restarting of the syscall (as if a tracer application,
like gdb, would never hijack a thread), then never did the subtracting
again. Fortunately, it doesn't assume that. Instead, it relies on the
contents of the register used to return values (%rax on x86_64), to
determine, again and again, if the syscall needs restarting, for
instance, on kernel's arch/x86/kernel/signal.c:

  switch (syscall_get_error(current, regs)) {
    case -ERESTARTNOHAND:
    case -ERESTARTSYS:
    case -ERESTARTNOINTR:
      regs->ax = regs->orig_ax;
      regs->ip -= 2;
      break;

where syscall_get_error(current, regs) returns the contents of %rax.

This block of code gets executed before the kernel sends the thread back
to userspace, but only after an attached tracer program has had a chance
to play with the registers. In libpulp's case, that means that, after
coming back from a live patching routine, the contents of %rax will have
the result from that routine. However, before detaching from the thread,
libpulp restores the context to the same it had when attaching,
including %rax,, which re-triggers the subtraction mechanism, as
exemplified below:

  1. Start the target program

    $ LD_PRELOAD=lib/.libs/libpulp.so tests/syscall_restart &
    [1] 1729

    (now the program is waiting for input (read syscall))

  2. Attaching to it (with a modified version of ulp_check that prints
     extra info) would reveal that the contents of %rax are -512
     (-ERESTARTSYS):

    ulp:   context.rax      : fffffffffffffe00

  3. After executing live patching routines, the contents of %rax
     represent the result from them (in the case of ulp_check, whether
     the live patch has benn applied or not):

    ulp:   curr.context.rax      : 0

  4. Before detaching, ulp_check restores all register:

    ulp:   save.context.rax      : fffffffffffffe00

  5. After detaching, the kernel notices that %rax equals -512
     (-512 == 0xfffffffffffffe00 == -ERESTARTSYS) and restarts the
     syscall (more precisely, it subtracts one instruction from the
     program counter, again). If %rax had not been restored, the syscall
     would not be restarted.

This patch adds a test case that tests that the read syscall actually
gets restarted after libpulp applies a live patch. It doesn't reveal any
problems with libpulp.