Make tools requirements mandatory
The commit

commit 15e9bba2e275dc59c149ab1eab9b2e4ef8e4540a
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Mon Apr 6 14:48:10 2020 -0300

    Check for headers required to build the tools

added checks for the presence of binutils' and libelf's development
files, but made them optional, in the hope that building just the
library would be useful. However, when any of those tools are missing,
make check will not run the test suite and exit without errors, which
is misleading. This patch makes these requirements mandatory in the
configure step.