Restore previous memory protection bits after live patching
When a live patch is applied, the nop padding area at the start of live
patchable functions gets modified. This requires that the access
protections for the affected memory pages have the writable bit set.
Before this patch, libpulp would set the protection for affected memory
pages to writable (PROT_WRITE), apply the live patch, then set the
protection to readable and executable (PROT_READ | PROT_EXEC),
regardless of the previous protection configuration. With this patch,
libpulp restores the protection bits to their former configuration.