Repurpose padding nops and drop binutils patch
Libpulp depends on a patched version of binutils to create the .ulp
section, which is later filled with trampolines for exported functions.
However, the trampolines need not be in a separate section. This patch
moves the trampolines from the .ulp section to the nop padding area
preceding each function built with -fpatchable-function-entry.

After library post-processing, the preceding nops area now looks like
the following diagram:

  +-------------------------+----------+---------------+---------------+
  | Trampoline:             |          |               |               |
  |   lea    off(%rip),%r11 | 16 bytes |               |               |
  |   push   %r11           |          |               |               |
  |   jmpq   <__ulp_entry>  |          |               |               |
  +-------------------------+----------+ PRE_NOPS_LEN  |               |
  | Detour (22 bytes):      |          |  (38 bytes)   |               |
  |   nop                   | 22 bytes |               | ULP_NOPS_LEN  |
  |   nop                   |          |               |  (40 bytes)   |
  |   ...                   |          |               |               |
  +-------------------------+----------+---------------+               |
  | Function entry:         |          2 bytes         |               |
  |   xchg %rax,%rax (nop)  |                          |               |
  +-------------------------+--------------------------+---------------+

i.e. the nop padding area needs to be 16 bytes longer than it used to
be, thus, gcc is now called with -fpatchable-function-entry=40,38.