Sanity check for __patchable_function_entries
When the argument to ulp_post lacks the __patchable_function_entries
section, ulp_post crashes. The absence of the section might happen when
the binary has not been built with -fpatchable-function-entry, or due to
GCC's PR 98971, when -flto is in use. This patch adds a sanity check to
detect either of these problems early with a friendlier error message.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=98971