Add a coding style and improve the README file
Currently, there is no consistent coding style across the code base,
which will probably lead to even more inconsistencies in the future. On
the one hand, fixing the coding style introduces lots of white space
changes, which makes it harder to revert or cherry-pick patches; on the
other hand, there are no libpulp releases so far, so settling on a
coding style now is probably harmless.

In order to avoid manual work, as well as to prevent further deviations
from some standard, this patch adds a .clang-format file, which is
supposed to be used with 'clang-format -style=file <filename>'. The
.clang-format file is largely based on the coding style used in the GNU
project.

A subsequent patch will apply the changes to the whole code base, with:

  find -name "*.[c|h]" -exec clang-format -style=file -i {} \;

The other changes in this patch rewrite some lines that did not play
well with clang-format.