Fix trigger retries
Right after hijacking all the threads from a target process, the trigger
tool tries to detect whether it would be safe to apply the live patch or
if the process could deadlock (due to malloc and dlopen locks). When it
is not safe, the trigger tool detaches from the process and attaches
again, 100 times, because doing so might lead to a safe condition.
However, the logic to decrement the attempts counter was misplaced and
led to 100 retries even when there was no deadlocking hazard, for
instance, if live patch application failed due to the metadata referred
to a non-existing function.