Interpose malloc and libdl functions to avoid the deadlock
As demonstrated by the test suite (deadlock.py), applying a live patch
might send the target process into a fatal deadlock. The deadlock
happens due to locks in the implementations of the malloc and dlopen
families of functions. In order to workaround the problem, libpulp
provided a patch for glibc, which exposed internal details of these
implementations.

Now, instead of requiring a patched glibc, libpulp interposes all
functions in the malloc and dlopen implementations, and keeps track of
their use. If any of the functions is in use when the application of a
live patch is triggered, libpulp steps back and asks the user to retry.
As a matter of fact, libpulp itself retries many times before it
actually gives up, because simply detaching and attaching again has a
high chance of unblocking the live patch application (except for
programs that make very heavy use of malloc and dlfcn functions).

Since the new solution does not have access to internal details of the
glibc implementation, it is not as efficient in finding a suitable
moment to apply a live patch, thus the number of threads in the test
case was reduced, otherwise the live patch is never applied. The new
amount of threads (15) is a magic number that could expose the problem
(if one of the functions, for instance calloc, is not interposed) but
succeeds on the machine the tests were executed.