Extend the CI to test for coding style problems
The CI now complains when a change adds coding style problems. It checks
the whole code base (all .c and .h files), not just changed-lines, which
is not ideal, but easier to get done since all files currently comply
with .clang-format. The check is informational only, and failing does
not block commits.

Apart from that, the test suite now only prints the logs of failing
tests, as well as it only prints them if make check failed (previously,
it tried to print the logs even if make check was not executed due to
errors in previous steps).