Whitelist __libc_start_main
When a C program is compiled with the GNU Toolchain in the usual
fashion, its entry point gets assigned to the _start function (provided
by glibc's crt0.o). After some basic initialization, _start calls
__libc_start_main (provided by glibc's libc.so), which initializes more
stuff, then finally calls the main function.

For the x86_64 architecture (and probably for all architectures), _start
is manually written in assembly, which means that it is not affected by
-fpatchable-function-entry, which further means that the Post tool does
not touch it at all. On the other hand, __libc_start_main is typically
written in C, thus affected by -fpatchable-function-entry, thus post
processed by the Post tool so that calls to it go through __ulp_entry.

Therefore, during process startup, when __libc_start_main gets called,
the library entrance tracking routine (__ulp_entry) kicks in and marks
libc.so as entered. Finally, since __libc_start_main calls main before
returning, the process is led to believe that libc.so is forever
entered, which is ultimately true, but useless from a live patching
point-of-view.

This patch whitelists __libc_start_main, so that it doesn't go through
__ulp_entry, which lets libc.so be live patched.